Fix dependency cycle lint error
