+++ src/modules/crud.js
@@ -1,7 +1,4 @@
-/* eslint-disable import/no-cycle */
-import ul from '../index.js';
-
-// Using local storage
+// Local storage functions
 export const getLocalStorage = () => {
   let tasks;
   if (localStorage.getItem('tasks') === null) {
@@ -10,30 +7,6 @@ export const getLocalStorage = () => {
     tasks = JSON.parse(localStorage.getItem('tasks'));
   }
   return tasks;
-};
-
-// Ui's
-export const addTaskToList = (task) => {
-  const taskItem = document.createElement('li');
-  let checkmark;
-  let completedClass;
-  if (task.completed === true) {
-    checkmark = '<span class="material-icons checkmark">done</span>';
-    completedClass = 'completed';
-  }
-  taskItem.classList.add('todo');
-  taskItem.innerHTML = `
-  <button type="button" class="checkbox ${completedClass}" id=${task.index}>${checkmark}</button>
-  <p contentEditable="true" class="desc">${task.description}</p>
-  <div class="dots">
-    <span class="material-icons">more_vert</span>
-  </div>
-  <div class="bin">
-    <span class="material-symbols-outlined delete-bin">delete</span>
-  </div>
-  `;
-
-  ul.appendChild(taskItem);
 };
 
 export const addNewTask = (task) => {
@@ -64,11 +37,7 @@ export const editTask = (newtext, index) => {
   }
 };
 
-export const displayTasks = () => {
-  const tasks = getLocalStorage();
-  tasks.forEach((task) => addTaskToList(task));
-};
-
+// Interactive functions
 export const markCompleted = (index) => {
   const tasks = getLocalStorage();
   tasks[index].completed = true;
@@ -79,4 +48,8 @@ export const markUnCompleted = (index) => {
   const tasks = getLocalStorage();
   tasks[index].completed = false;
   localStorage.setItem('tasks', JSON.stringify(tasks));
+};
+
+export const clearCompleted = () => {
+
 };
